Update cosmors_calculation.ipynb
When using RDKit 2025.03.5, the original function for counting atoms within a ring does not work as expected, since Chem.MolFromXYZFile() fails to correctly interpret the XYZ file.  As a result, the variable num_ring_atoms is always returned as zero, meaning that the ring correction is not applied in the solvation energy calculation. The update enables the code to correctly reconstruct the molecular bonds, allowing for accurate ring detection.